requirements.txt
python-dotenv
google-generativeai
requests
flask
discord.py
gunicorn